Update README.md
added an elevator pitch, listed some key features, a general description (same as elevator pitch), included a basic description for most of the technologies that will be used, and included some .png files which are photos drawings I did that show the basic layout of the pages. The page layout is subject to change.